# Optimize sdb_array_add_num: always use SDB_NUM_BASE and don't check for decimal