https://github.com/radare/sdb/pull/156/files

```
# t1 is an empty project

# Before

% time (repeat 10 ~/Dev/Bin/radare2/release/binr/radare2/radare2 -p t1 -c '' -Q a.out)
( repeat 10; do; ~/Dev/Bin/radare2/release/binr/radare2/radare2 -p t1 -c '' -)  1.21s user 0.05s system 98% cpu 1.277 total

# After

% time (repeat 10 ~/Dev/Bin/radare2/release/binr/radare2/radare2 -p t1 -c '' -Q a.out)
( repeat 10; do; ~/Dev/Bin/radare2/release/binr/radare2/radare2 -p t1 -c '' -)  0.62s user 0.08s system 96% cpu 0.729 total
```

sdb can be optimized a lot and become more readable if you use some coverage tool (with radare2-regressions tests) and remove redundant code (it might be risky to remove the check here).