@MaskRay 2018-08-19
BTW, I don't understand what

```c
// line 29
#define META_RANGE_BASE(x) ((x)>>12)
#define META_RANGE_SIZE 0xfff
	for (; base < base2; base += META_RANGE_SIZE) { // 0xfff not 0x1000 ???
		const char *key = sdb_fmt ("range.0x%"PFMT64x, base);
```

means